Add Data Validation and  Query to StockRepo/Controller

+++ Controllers/CommentController.cs
@@ -22,13 +22,17 @@ public class CommentsController : ControllerBase
     [HttpGet]
     public async Task<IActionResult> GetAll()
     {
+        if(!ModelState.IsValid)
+            return BadRequest(ModelState);
         var comments = await _commentRepo.GetAllAsync();
         return Ok(comments);
     }
 
-    [HttpGet("{id}")]
-    public async Task<IActionResult> GetComment(int id)
+    [HttpGet("{id}:int")]
+    public async Task<IActionResult> GetById(int id)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState);
         var comment = await _commentRepo.GetByIdAsync(id);
         if (comment == null)
             return NotFound();
@@ -36,25 +40,26 @@ public class CommentsController : ControllerBase
         return Ok(comment.ToCommentDto());
     }
 
-    [HttpPost("{stockId}")]
+    [HttpPost("{stockId}: int")]
     public async Task<IActionResult> Create([FromRoute] int stockId, [FromBody] CreateCommentDto commentDto)
     {
-       
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState);
         if (!await _stockRepo.StockExists(stockId))
         {
             return BadRequest("Stock does not exist.");
         }
-            
-
         var commentModel = commentDto.ToCommentFromCreate(stockId);
         await _commentRepo.CreateAsync(commentModel);
-        return CreatedAtAction(nameof(GetComment),new { id = commentModel.Id},commentModel.ToCommentDto()
+        return CreatedAtAction(nameof(GetById),new { id = commentModel.Id},commentModel.ToCommentDto()
         );
     }
     [HttpPut]
-    [Route("{id}")]
+    [Route("{id}:int")]
     public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState);
         var comment = await _commentRepo.UpdateAsync(id, updateDto.ToCommentFromUpdate(id));
         if (comment == null)
         {
@@ -63,9 +68,11 @@ public class CommentsController : ControllerBase
         return Ok(comment.ToCommentDto());
     }
     [HttpDelete]
-    [Route("{id}")]
+    [Route("{id}:int")]
     public async Task<IActionResult> Delete([FromRoute] int id)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState);
         var comment = await _commentRepo.DeleteAsync(id);
         if (comment == null)
         {
@@ -75,3 +82,4 @@ public class CommentsController : ControllerBase
     }
 
 }
+
